Change protocol fee from denominator to bips

+++ src/test/ProtocolFeeControllerTest.sol
@@ -9,28 +9,28 @@ import {PoolKey} from "../types/PoolKey.sol";
 contract ProtocolFeeControllerTest is IProtocolFeeController {
     using PoolIdLibrary for PoolKey;
 
-    mapping(PoolId => uint16) public swapFeeForPool;
+    mapping(PoolId => uint32) public swapFeeForPool;
 
-    function protocolFeeForPool(PoolKey memory key) external view returns (uint16) {
+    function protocolFeeForPool(PoolKey memory key) external view returns (uint32) {
         return swapFeeForPool[key.toId()];
     }
 
     // for tests to set pool protocol fees
-    function setSwapFeeForPool(PoolId id, uint16 fee) external {
+    function setSwapFeeForPool(PoolId id, uint32 fee) external {
         swapFeeForPool[id] = fee;
     }
 }
 
 /// @notice Reverts on call
 contract RevertingProtocolFeeControllerTest is IProtocolFeeController {
-    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint16) {
+    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint32) {
         revert();
     }
 }
 
 /// @notice Returns an out of bounds protocol fee
 contract OutOfBoundsProtocolFeeControllerTest is IProtocolFeeController {
-    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint16) {
+    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint32) {
         // set both swap and withdraw fees to 1, which is less than MIN_PROTOCOL_FEE_DENOMINATOR
         return 0x001001;
     }
@@ -38,7 +38,7 @@ contract OutOfBoundsProtocolFeeControllerTest is IProtocolFeeController {
 
 /// @notice Return a value that overflows a uint24
 contract OverflowProtocolFeeControllerTest is IProtocolFeeController {
-    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint16) {
+    function protocolFeeForPool(PoolKey memory /* key */ ) external pure returns (uint32) {
         assembly {
             let ptr := mload(0x40)
             mstore(ptr, 0xFFFFAAA001)
@@ -49,7 +49,7 @@ contract OverflowProtocolFeeControllerTest is IProtocolFeeController {
 
 /// @notice Returns data that is larger than a word
 contract InvalidReturnSizeProtocolFeeControllerTest is IProtocolFeeController {
-    function protocolFeeForPool(PoolKey memory /* key */ ) external view returns (uint16) {
+    function protocolFeeForPool(PoolKey memory /* key */ ) external view returns (uint32) {
         address a = address(this);
         assembly {
             let ptr := mload(0x40)
